Add GenDispatch.implementsIsSerializable
It is useful for the generated Action and Result pairs to be able
to implement IsSerializable, in addition to inheriting the Serializable
implementation from the top-level classes. One use case is being able
to skip generating a serialization policy.

Currently this is only possible by declaring base action and result
classes, which is cumbersome and overkill for a simple marker
interface implementation.